Refactored feature detector to make more generic
