+++ source/LibMultiSense/details/channel.cc
@@ -79,7 +79,6 @@ impl::impl(const std::string& address, const RemoteHeadChannel& cameraId, const 
     m_rxLargeBufferPool(),
     m_rxSmallBufferPool(),
     m_imageMetaCache(IMAGE_META_CACHE_DEPTH),
-    // m_featureDetectorMetaCache(FEATURE_DETECTOR_META_CACHE_DEPTH),
     m_secondaryAppMetaCache(SECONDARY_APP_META_CACHE_DEPTH),
     m_udpAssemblerMap(),
     m_dispatchLock(),
@@ -94,7 +93,6 @@ impl::impl(const std::string& address, const RemoteHeadChannel& cameraId, const 
     m_imuListeners(),
     m_compressedImageListeners(),
     m_secondaryAppListeners(),
-    // m_featureDetectorListeners(),
     m_watch(),
     m_messages(),
     m_streamsEnabled(0),
@@ -271,11 +269,6 @@ void impl::cleanup()
         its != m_secondaryAppListeners.end();
         its ++)
         delete *its;
-    // std::list<FeatureDetectorListener*>::const_iterator itf;
-    // for(itf  = m_featureDetectorListeners.begin();
-    //     itf != m_featureDetectorListeners.end();
-    //     ++ itf)
-    //     delete *itf;
     BufferPool::const_iterator it;
     for(it  = m_rxLargeBufferPool.begin();
         it != m_rxLargeBufferPool.end();
@@ -292,7 +285,6 @@ void impl::cleanup()
     m_imuListeners.clear();
     m_secondaryAppListeners.clear();
     m_compressedImageListeners.clear();
-    // m_featureDetectorListeners.clear();
     m_rxLargeBufferPool.clear();
     m_rxSmallBufferPool.clear();
 
@@ -481,12 +473,6 @@ wire::SourceType impl::sourceApiToWire(DataSource mask)
     if (mask & Source_Disparity_Cost)                   wire_mask |= wire::SOURCE_DISPARITY_COST;
     if (mask & Source_Jpeg_Left)                        wire_mask |= wire::SOURCE_JPEG_LEFT;
     if (mask & Source_Rgb_Left)                         wire_mask |= wire::SOURCE_RGB_LEFT;
-    if (mask & Source_Feature_Left)                     wire_mask |= wire::SOURCE_FEATURE_LEFT;
-    if (mask & Source_Feature_Right)                    wire_mask |= wire::SOURCE_FEATURE_RIGHT;
-    if (mask & Source_Feature_Aux)                      wire_mask |= wire::SOURCE_FEATURE_AUX;
-    if (mask & Source_Feature_Rectified_Left)           wire_mask |= wire::SOURCE_FEATURE_RECTIFIED_LEFT;
-    if (mask & Source_Feature_Rectified_Right)          wire_mask |= wire::SOURCE_FEATURE_RECTIFIED_RIGHT;
-    if (mask & Source_Feature_Rectified_Aux)            wire_mask |= wire::SOURCE_FEATURE_RECTIFIED_AUX;
     if (mask & Source_Lidar_Scan)                       wire_mask |= wire::SOURCE_LIDAR_SCAN;
     if (mask & Source_Imu)                              wire_mask |= wire::SOURCE_IMU;
     if (mask & Source_Pps)                              wire_mask |= wire::SOURCE_PPS;
@@ -532,12 +518,6 @@ DataSource impl::sourceWireToApi(wire::SourceType mask)
     if (mask & wire::SOURCE_DISPARITY_COST)                 api_mask |= Source_Disparity_Cost;
     if (mask & wire::SOURCE_JPEG_LEFT)                      api_mask |= Source_Jpeg_Left;
     if (mask & wire::SOURCE_RGB_LEFT)                       api_mask |= Source_Rgb_Left;
-    if (mask & wire::SOURCE_FEATURE_LEFT)                   api_mask |= Source_Feature_Left;
-    if (mask & wire::SOURCE_FEATURE_RIGHT)                  api_mask |= Source_Feature_Right;
-    if (mask & wire::SOURCE_FEATURE_AUX)                    api_mask |= Source_Feature_Aux;
-    if (mask & wire::SOURCE_FEATURE_RECTIFIED_LEFT)         api_mask |= Source_Feature_Rectified_Left;
-    if (mask & wire::SOURCE_FEATURE_RECTIFIED_RIGHT)        api_mask |= Source_Feature_Rectified_Right;
-    if (mask & wire::SOURCE_FEATURE_RECTIFIED_AUX)          api_mask |= Source_Feature_Rectified_Aux;
     if (mask & wire::SOURCE_LIDAR_SCAN)                     api_mask |= Source_Lidar_Scan;
     if (mask & wire::SOURCE_IMU)                            api_mask |= Source_Imu;
     if (mask & wire::SOURCE_PPS)                            api_mask |= Source_Pps;
