+++ source/LibMultiSense/include/MultiSense/MultiSenseChannel.hh
@@ -377,23 +377,6 @@ public:
                                        void         *userDataP=NULL) = 0;
 
     /**
-     * Add a user defined callback attached to the feature detector stream.
-     *
-     * Each callback will create a unique internal thread
-     * dedicated to servicing the callback.
-     *
-     * @param callback A user defined feature_detector::Callback to send Ground
-     * Surface data to
-     *
-     * @param userDataP A pointer to arbitrary user data.
-     *
-     * @return A crl::multisense::Status indicating if the callback registration
-     * succeeded or failed
-     */
-    // virtual Status addIsolatedCallback(feature_detector::Callback callback,
-                                       // void         *userDataP=NULL) = 0;
-
-    /**
      * Unregister a user defined image::Callback. This stops the callback
      * from receiving image data.
      *
@@ -478,18 +461,6 @@ public:
     virtual Status removeIsolatedCallback(apriltag::Callback callback) = 0;
 
     /**
-     * Unregister a user defined feature_detector::Callback. This stops the callback
-     * from receiving feature data
-     *
-     * @param callback The user defined feature_detector::Callback to unregister
-     *
-     * @return A crl::multisense::Status indicating if the callback deregistration
-     * succeeded or failed
-     */
-
-    // virtual Status removeIsolatedCallback(feature_detector::Callback callback) = 0;
-
-    /**
      * Unregister a user defined secondary_app::Callback. This stops the callback
      * from receiving ground surface data
      *
@@ -499,19 +470,6 @@ public:
      * succeeded or failed
      */
     virtual Status removeIsolatedCallback(secondary_app::Callback callback) = 0;
-
-    /**
-     * Provides an interface to convert generic secondary app data to usable feature detector data
-     *
-     * @param h the application defined feature detector header
-     * @param data the generic payload from the secondary application
-     * @param len the length of the generic payload
-     * @param frameId the frameId for metadata lookup
-     *
-     * @return A crl::multisense::Status indicating if the callback deregistration
-     * succeeded or failed
-     */
-    virtual Status secondaryAppDataExtract(feature_detector::Header &h, const secondary_app::Header &origh) = 0;
 
     /**
      * Reserve image or lidar data within a isolated callback so it is available
@@ -1112,26 +1070,6 @@ public:
     virtual Status setApriltagParams (const system::ApriltagParams& params) = 0;
 
     /**
-     * Set the feature detector config associated with the MultiSense device
-     *
-     * @param params The feature detector parameters to send to the on-camera feature detector
-     *               application
-     *
-     * @return A crl::multisense::Status indicating if the params were successfully set
-     */
-    virtual Status setFeatureDetectorConfig (const system::FeatureDetectorConfig& params) = 0;
-
-    /**
-     * Get the feature detector parameters associated with the MultiSense device
-     *
-     * @param params The feature detector parameters to send to the on-camera feature detector
-     *               application
-     *
-     * @return A crl::multisense::Status indicating if the params were successfully set
-     */
-    virtual Status getFeatureDetectorConfig (system::FeatureDetectorConfig& params) = 0;
-
-    /**
      * Get the secondary application config parameters associated with the MultiSense device
      *
      * @param params The secondary application parameters to send to the secondary application
@@ -1148,7 +1086,7 @@ public:
      *
      * @return A crl::multisense::Status indicating if the params were successfully set
      */
-    virtual Status setSecondaryAppConfig (const system::SecondaryAppConfig & c) = 0;
+    virtual Status setSecondaryAppConfig (system::SecondaryAppConfig & c) = 0;
 
     /**
      * Query the secondary application(s) registered with the MultiSense device
