+++ source/LibMultiSense/include/MultiSense/details/listeners.hh
@@ -235,7 +235,6 @@ typedef Listener<imu::Header,              imu::Callback>              ImuListen
 typedef Listener<compressed_image::Header, compressed_image::Callback> CompressedImageListener;
 typedef Listener<ground_surface::Header,   ground_surface::Callback>   GroundSurfaceSplineListener;
 typedef Listener<apriltag::Header,         apriltag::Callback>         AprilTagDetectionListener;
-typedef Listener<feature_detector::Header, feature_detector::Callback> FeatureDetectorListener;
 typedef Listener<secondary_app::Header,    secondary_app::Callback>    SecondaryAppListener;
 
 } // namespace details
